requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Processing
easyocr==1.7.0
PyMuPDF==1.23.9
Pillow==10.1.0
python-docx==1.1.0

# ML/AI
openai==1.3.8
sentence-transformers==2.2.2

# Cache
redis==5.0.1

# Monitoring
prometheus-client==0.19.0

# UI
streamlit==1.28.1

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0

# LLM Integration
openai==1.3.8
vllm==0.2.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Vector Database
pymilvus==2.3.4
milvus==2.3.4

# Advanced Processing
langchain==0.0.350
langchain-community==0.0.1
tiktoken==0.5.2
